refactor(api): move prescriber-serializer
to team bounded context
