+++ api/src/shared/prescriber-management/application/prescriber-informations-controller.js
@@ -1,5 +1,5 @@
 import { usecases } from '../domain/usecases/index.js';
-import * as prescriberSerializer from '../infrastructure/serializers/jsonapi/prescriber-serializer.js';
+import * as prescriberSerializer from '../../../team/infrastructure/serializers/jsonapi/prescriber-serializer.js';
 
 const get = async function (request, h, dependencies = { prescriberSerializer }) {
   const authenticatedUserId = request.auth.credentials.userId;
